requirements.txt
docker
pyyaml
jinja2
